+++ index.js
@@ -6,4 +6,4 @@ export {uuid};
 export {changeFavicon} from './src/changeFavicon';
 export {getBestLocale, getBrowserLocales} from './src/locales';
 export {URI};
-export {sanitizeString} from './src/sanitizeString';
+export {sanitizeFileName} from './src/sanitizeFileName';